feat: 新增1、客户批量删除/单个删除，2、客户手动移入公海，3、公海管理列表，4、销售领取客户

+++ src/main/java/com/crm/entity/Customer.java
@@ -14,10 +14,10 @@ import lombok.Setter;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
- * @author vact
+ * @author crm
  * @since 2025-10-12
  */
 @Getter
@@ -44,11 +44,11 @@ public class Customer {
 
     @ApiModelProperty("客户级别 0-普通客户，1-优质客户，2-重点客户，3-其他")
     @TableField("level")
-    private Byte level;
+    private Integer level;
 
     @ApiModelProperty("客户来源 0-个人资源，1-客户介绍，2-官网咨询，3-公众号，4-抖音，5-小红书，6-其他")
     @TableField("source")
-    private Byte source;
+    private Integer source;
 
     @ApiModelProperty("客户地址")
     @TableField("address")
@@ -56,7 +56,7 @@ public class Customer {
 
     @ApiModelProperty("跟进状态 0-新客，1-待再次沟通，2-有意向，3-转入商机，4-无效")
     @TableField("follow_status")
-    private Byte followStatus;
+    private Integer followStatus;
 
     @ApiModelProperty("下次跟进时间")
     @TableField("next_follow_status")
@@ -72,7 +72,7 @@ public class Customer {
 
     @ApiModelProperty("是否转入公海 0-未转入公海，1-已转入公海")
     @TableField("is_public")
-    private Byte isPublic;
+    private Integer isPublic;
 
     @ApiModelProperty("客户所属的员工id")
     @TableField("owner_id")
@@ -80,11 +80,11 @@ public class Customer {
 
     @ApiModelProperty("是否为关键决策人 0-是，1-否")
     @TableField("is_key_decision_maker")
-    private Byte isKeyDecisionMaker;
+    private Integer isKeyDecisionMaker;
 
     @ApiModelProperty("性别 0-男，1-女 2-保密")
     @TableField("gender")
-    private Byte gender;
+    private Integer gender;
 
     @ApiModelProperty("成交次数")
     @TableField("deal_count")
@@ -93,7 +93,7 @@ public class Customer {
     @ApiModelProperty("逻辑删除")
     @TableField(value = "delete_flag", fill = FieldFill.INSERT)
     @TableLogic
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 
     @ApiModelProperty("创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
